Run tenant data injector for new tenants - WIP
