--- .devcontainer/devcontainer.json
@@ -1,27 +1,0 @@
-{
-	"name": "Ubuntu",
-	"image": "mcr.microsoft.com/devcontainers/base:0-focal",
-
-	// Use 'forwardPorts' to make a list of ports inside the container available locally.
-	"forwardPorts": [3478],
-
-	// Use 'postCreateCommand' to run commands after the container is created.
-	"postCreateCommand": "curl http://openline.sh/install.sh | sh",
-	"postStartCommand": "docker run -d --network=host coturn/coturn -v -z -n -X $(hostname -I|cut -f1 -d ' ') -L 0.0.0.0 --min-port=10000 --max-port=20000",
-	// Set `remoteUser` to `root` to connect as root instead. More info: https://aka.ms/vscode-remote/containers/non-root.
-	"remoteUser": "vscode",
-	"features": {
-		"ghcr.io/devcontainers/features/github-cli:1": {},
-		"ghcr.io/devcontainers/features/kubectl-helm-minikube:1": {},
-		"ghcr.io/devcontainers/features/docker-in-docker:1": {},
-		"ghcr.io/devcontainers/features/go:1": {},
-		"ghcr.io/devcontainers/features/node:1": {}
-	},
-	"customizations": {
-		"vscode": {
-			"extensions": [
-				"ms-kubernetes-tools.vscode-kubernetes-tools"
-			]
-		}
-	}
-}
